Restoring stick input logic
I had a misconception as to how the stick input logic related to power value generation

+++ src/Drive/Drive.h
@@ -83,7 +83,7 @@ class Drive {
     // MotorControl* M1;
     // MotorControl* M2;
     MotorControl M1, M2;
-    int stickForwardRev, stickTurn;
+    float stickForwardRev, stickTurn;
     float lastTurnPwr;
     float turnPower;
